// v2-src/types.ts
export interface OperatorParameterMetadata {
  name: string
  description: string
  aliases: string[]
  required: boolean
  type: ExpectedType
  default?: unknown
}

export interface FragmentParameterMetadata {
  name: string
  type: string | string[]
  required: boolean
  description?: string
  default?: unknown
}

export interface OperatorData {
  description: string
  aliases: string[]
  parameters: OperatorParameterMetadata[]
}